backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
pyyaml>=6.0.1
kubernetes>=28.1.0
prometheus-client>=0.19.0
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
websockets>=12.0
httpx>=0.25.2
black>=23.11.0
mypy>=1.7.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
hypothesis>=6.92.1